Refactor HL7 Transformer to hl7_phw_transformer
