+++ hl7_phw_transformer/pyproject.toml
@@ -1,5 +1,5 @@
 [project]
-name = "hl7-transformer"
+name = "hl7-phw-transformer"
 version = "0.1.0"
 description = "HL7v2 transformer"
 requires-python = ">=3.13"
