+++ pipeline-ado/hl7phwtransformer-build.yml
@@ -5,7 +5,7 @@ trigger:
       - main
   paths:
     include:
-      - hl7_transformer/*
+      - hl7_phw_transformer/*
       - shared_libs/*
       - pipeline-ado/templates/*
       - ca-certs/*
@@ -37,7 +37,7 @@ stages:
     jobs:
       - template: templates/code-quality-template.yml
         parameters:
-          appName: 'hl7_transformer'
+          appName: 'hl7_phw_transformer'
           appDisplayName: 'HL7 Transformer'
           pythonVersion: '3.13'
           poolName: ${{ parameters.POOL_SELECTION }}
@@ -50,9 +50,9 @@ stages:
         parameters:
           jobName: 'BuildPushHL7Transformer'
           displayName: 'Build and Push HL7 Transformer'
-          appName: 'hl7transformer'
-          dockerfilePath: 'hl7_transformer/Dockerfile'
-          buildContext: 'hl7_transformer'
+          appName: 'hl7phwtransformer'
+          dockerfilePath: 'hl7_phw_transformer/Dockerfile'
+          buildContext: 'hl7_phw_transformer'
           additionalContexts: "shared_libs=shared_libs,ca-certs=ca-certs"
           acrName: $(acrName)
           azureServiceConnection: $(azureServiceConnection)
