+++ pipeline-ado/release-apps.yml
@@ -120,7 +120,7 @@ stages:
         parameters:
           jobName: 'ReleaseAppHL7TransformerDev'
           displayName: 'Release App HL7Transformer - Dev'
-          appName: 'hl7transformer'
+          appName: 'hl7phwtransformer'
           appFunctionMidfix: ''
           resourceGroupName: 'UK-South-DHCW-IntHub-DEV-RG'
           azureLocation: $(azureLocation)
@@ -312,7 +312,7 @@ stages:
         parameters:
           jobName: 'ReleaseAppHL7TransformerTest'
           displayName: 'Release App HL7Transformer - Test'
-          appName: 'hl7transformer'
+          appName: 'hl7phwtransformer'
           appFunctionMidfix: ''
           resourceGroupName: 'UK-South-DHCW-IntHub-TST-RG'
           azureLocation: $(azureLocation)
@@ -493,7 +493,7 @@ stages:
         parameters:
           jobName: 'ReleaseAppHL7TransformerPreProd'
           displayName: 'Release App HL7Transformer - Pre-Prod'
-          appName: 'hl7transformer'
+          appName: 'hl7phwtransformer'
           appFunctionMidfix: ''
           resourceGroupName: 'UK-South-DHCW-IntHub-PPD-RG'
           azureLocation: $(azureLocation)
@@ -674,7 +674,7 @@ stages:
         parameters:
           jobName: 'ReleaseAppHL7TransformerProd'
           displayName: 'Release App HL7Transformer - Production'
-          appName: 'hl7transformer'
+          appName: 'hl7phwtransformer'
           appFunctionMidfix: ''
           resourceGroupName: 'UK-South-DHCW-IntHub-PRD-RG'
           azureLocation: $(azureLocation)
